# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/tasks": {
      "post": {
        "description": "Cria uma nova tarefa associando-a ao usuário autenticado como proprietário.",
        "operationId": "TasksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tarefa criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos ou data limite no passado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar nova tarefa",
        "tags": [
          "tasks"
        ]
      },
      "get": {
        "description": "Lista tarefas com paginação, busca por texto, filtro por status/prioridade/categoria e ordenação.",
        "operationId": "TasksController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número da página",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Itens por página",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Termo de busca",
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filtro por status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "IN_PROGRESS",
                "COMPLETED",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "description": "Filtro por prioridade",
            "schema": {
              "type": "string",
              "enum": [
                "LOW",
                "MEDIUM",
                "HIGH",
                "URGENT"
              ]
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filtro por categoria",
            "schema": {
              "type": "string",
              "enum": [
                "PERSONAL",
                "WORK",
                "STUDY",
                "SHOPPING",
                "HEALTH",
                "OTHER"
              ]
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Campo de ordenação",
            "schema": {
              "default": "createdAt",
              "type": "string",
              "enum": [
                "createdAt",
                "dueDate",
                "title",
                "priority",
                "status",
                "category"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Direção da ordenação",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de tarefas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTasksResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar tarefas paginadas",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/tasks/{id}": {
      "get": {
        "description": "Retorna os detalhes de uma tarefa específica pertencente ao usuário (ou qualquer tarefa para ADMIN).",
        "operationId": "TasksController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados da tarefa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado (tarefa de outro proprietário)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Tarefa não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter detalhes de uma tarefa",
        "tags": [
          "tasks"
        ]
      },
      "put": {
        "description": "Atualiza os dados ou status de uma tarefa respeitando as regras de negócio e ownership.",
        "operationId": "TasksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tarefa atualizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDto"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos ou alteração inválida em tarefa concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Tarefa não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar tarefa",
        "tags": [
          "tasks"
        ]
      },
      "delete": {
        "description": "Realiza a remoção lógica (soft delete) da tarefa.",
        "operationId": "TasksController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Tarefa excluída com sucesso"
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Acesso negado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Tarefa não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Excluir tarefa (remoção lógica)",
        "tags": [
          "tasks"
        ]
      }
    }
  },
  "info": {
    "title": "AppStart API",
    "description": "API REST do AppStart - Especificação OpenAPI 3.0",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "appstart_session",
        "description": "Cookie de sessão local opaca"
      }
    },
    "schemas": {
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "https://appstart.example/problems/validation-error"
          },
          "title": {
            "type": "string",
            "example": "Dados inválidos"
          },
          "status": {
            "type": "number",
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "Um ou mais campos precisam ser corrigidos."
          },
          "instance": {
            "type": "string",
            "example": "/api/v1/users"
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "requestId": {
            "type": "string",
            "example": "01J..."
          },
          "timestamp": {
            "type": "string",
            "example": "2026-08-31T20:00:00.000Z"
          },
          "errors": {
            "type": "object",
            "example": {
              "email": [
                "Informe um e-mail válido."
              ]
            },
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "timestamp"
        ]
      },
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 100
          },
          "totalPages": {
            "type": "number",
            "example": 5
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
...
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Título da tarefa",
            "example": "Configurar monitoramento de produção",
            "minLength": 3,
            "maxLength": 150
          },
          "description": {
            "type": "string",
            "description": "Descrição detalhada da tarefa",
            "example": "Definir alertas para uso de CPU e memória.",
            "maxLength": 1000
          },
          "priority": {
            "type": "string",
            "description": "Prioridade da tarefa",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "URGENT"
            ],
            "default": "MEDIUM"
          },
          "category": {
            "type": "string",
            "description": "Categoria da tarefa",
            "enum": [
              "PERSONAL",
              "WORK",
              "STUDY",
              "SHOPPING",
              "HEALTH",
              "OTHER"
            ],
            "default": "OTHER"
          },
          "dueDate": {
            "type": "string",
            "description": "Data de entrega limite (ISO 8601)",
            "example": "2026-12-31T23:59:59.000Z"
          }
        },
        "required": [
          "title"
        ]
      },
      "TaskOwnerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identificador único do proprietário",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
          },
          "name": {
            "type": "string",
            "description": "Nome do proprietário",
            "example": "Ada Lovelace"
          },
          "email": {
            "type": "string",
            "description": "E-mail do proprietário",
            "example": "ada@example.com"
          }
        },
        "required": [
          "id",
          "name",
          "email"
        ]
      },
      "TaskDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identificador único da tarefa",
            "example": "b0eebc99-9c0b-4ef8-bb6d-6bb9bd380a22"
          },
          "title": {
            "type": "string",
            "description": "Título da tarefa",
            "example": "Configurar monitoramento"
          },
          "description": {
            "type": "object",
            "description": "Descrição da tarefa",
            "nullable": true
          },
          "status": {
            "type": "string",
            "description": "Status da tarefa",
            "enum": [
              "PENDING",
              "IN_PROGRESS",
              "COMPLETED",
              "CANCELLED"
            ],
            "example": "PENDING"
          },
          "priority": {
            "type": "string",
            "description": "Prioridade da tarefa",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "URGENT"
            ],
            "example": "MEDIUM"
          },
          "category": {
            "type": "string",
            "description": "Categoria da tarefa",
            "enum": [
              "PERSONAL",
              "WORK",
              "STUDY",
              "SHOPPING",
              "HEALTH",
              "OTHER"
            ],
            "example": "OTHER"
          },
          "dueDate": {
            "type": "object",
            "description": "Data limite de entrega",
            "nullable": true
          },
          "ownerId": {
            "type": "string",
            "description": "Identificador do usuário proprietário",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
          },
          "owner": {
            "description": "Dados resumidos do proprietário",
            "allOf": [
              {
                "$ref": "#/components/schemas/TaskOwnerDto"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "description": "Data de criação"
          },
          "updatedAt": {
            "type": "string",
            "description": "Data de última atualização"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "priority",
          "category",
          "ownerId",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedTasksResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Lista de tarefas paginadas",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskDto"
            }
          },
          "meta": {
            "description": "Metadados de paginação",
            "allOf": [
              {
                "$ref": "#/components/schemas/PaginationMetaDto"
              }
            ]
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Título da tarefa",
            "example": "Configurar monitoramento avançado",
            "minLength": 3,
            "maxLength": 150
          },
          "description": {
            "type": "string",
            "description": "Descrição detalhada da tarefa",
            "maxLength": 1000
          },
          "status": {
            "type": "string",
            "description": "Status atual da tarefa",
            "enum": [
              "PENDING",
              "IN_PROGRESS",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "priority": {
            "type": "string",
            "description": "Prioridade da tarefa",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "URGENT"
            ]
          },
          "category": {
            "type": "string",
            "description": "Categoria da tarefa",
            "enum": [
              "PERSONAL",
              "WORK",
              "STUDY",
              "SHOPPING",
              "HEALTH",
              "OTHER"
            ]
          },
          "dueDate": {
            "type": "string",
            "description": "Data de entrega limite (ISO 8601)"
          }
        }
      }
    }
  }
}